requirements.txt
speechrecognition
pyaudio

newsapi-python==0.2.5
pyowm==2.10.0

git+ssh://git@cordelianew.university.innopolis.ru:2222/hri/speech-to-intents.git
git+ssh://git@cordelianew.university.innopolis.ru:2222/hri/pyrhvoice.git